# Metal: PP speedup