Speedup is achieved via
* Faster `soft_max`, `diag_mask_inf`, `scale`, `silu`, `gelu`
* A new kernel for `f16 x f32` matrix multiplications that can not be done via the usual matrix multiplication kernel because the tensors are not contiguous. This mostly affects the `K x Q` matrix multiplication, the importance of which grows with increasing context / batch size. 
* Tuning of the k_quants dequantization kernels. The effect is most strong for `Q5_K` and `Q6_K`. 

The table gives results on a 30-core M2 Max. TG performance is mostly unaffected, but there are some very minor gains here and there. For PP the lion share of the speedup comes from the new `f16 x f32` matrix multiplication kernel, which somehow does not benefit as much the Falcon model, so this is something left to look Into.

| model                          | backend    | test       |     t/s (Master) |       t/s (PR)   |  Speedup |
| ------------------------------ | ---------- | ---------- | ---------------: | ---------------: | -------: |
| Falcon 7B mostly F16           | Metal      | pp 512     |    365.11 ± 0.06 |    380.18 ± 0.13 |   1.041  |
| LLaMA 7B mostly F16            | Metal      | pp 512     |    489.73 ± 0.32 |    539.46 ± 0.31 |   1.102  |
| LLaMA 7B mostly Q8_0           | Metal      | pp 512     |    445.27 ± 0.13 |    486.91 ± 0.18 |   1.094  |
| LLaMA 7B mostly Q4_0           | Metal      | pp 512     |    445.65 ± 0.30 |    495.91 ± 0.38 |   1.113  |
| LLaMA 7B mostly Q4_1           | Metal      | pp 512     |    448.26 ± 0.11 |    494.31 ± 0.26 |   1.103  |
| LLaMA 7B mostly Q6_K           | Metal      | pp 512     |    367.58 ± 0.15 |    413.82 ± 0.39 |   1.126  |
| LLaMA 7B mostly Q5_K - Small   | Metal      | pp 512     |    366.29 ± 0.14 |    413.35 ± 0.27 |   1.128  |
| LLaMA 7B mostly Q4_K - Small   | Metal      | pp 512     |    399.22 ± 0.28 |    438.64 ± 0.23 |   1.099  |
| LLaMA 7B mostly Q3_K - Small   | Metal      | pp 512     |    379.32 ± 0.18 |    417.91 ± 0.25 |   1.102  |
| Falcon 7B mostly F16           | Metal      | tg 128     |     23.34 ± 0.05 |     23.33 ± 0.04 |   1.000  |
| LLaMA 7B mostly F16            | Metal      | tg 128     |     24.25 ± 0.10 |     24.26 ± 0.04 |   1.000  |
| LLaMA 7B mostly Q8_0           | Metal      | tg 128     |     40.09 ± 0.34 |     40.16 ± 0.27 |   1.000  |
| LLaMA 7B mostly Q4_0           | Metal      | tg 128     |     62.23 ± 0.88 |     62.53 ± 0.06 |   1.005  |
| LLaMA 7B mostly Q4_1           | Metal      | tg 128     |     57.93 ± 0.49 |     58.88 ± 0.78 |   1.016  |
| LLaMA 7B mostly Q6_K           | Metal      | tg 128     |     42.68 ± 0.54 |     42.93 ± 0.48 |   1.006  |
| LLaMA 7B mostly Q5_K - Small   | Metal      | tg 128     |     43.90 ± 0.65 |     44.67 ± 0.03 |   1.017  |
| LLaMA 7B mostly Q4_K - Small   | Metal      | tg 128     |     56.42 ± 0.98 |     57.31 ± 0.44 |   1.016  |
| LLaMA 7B mostly Q3_K - Small   | Metal      | tg 128     |     45.67 ± 0.57 |     46.19 ± 0.50 |   1.011  |