@ggerganov 2023-09-09
M2 Ultra results:

| model                          |       size | test       |    th |     master   t/s |      PR      t/s |    spedup      |
| ------------------------------ | ---------: | ---------- | ---------: | ---------------: | ---------------: | --------------: |
| LLaMA 7B mostly F16            |  12.55 GiB | pp 512     |          4 |   1125.99 ± 0.69 |   1265.30 ± 1.54 |  1.124          |
| LLaMA 7B mostly Q8_0           |   6.67 GiB | pp 512     |          4 |   1029.88 ± 0.27 |   1142.84 ± 0.84 |  1.110          |
| LLaMA 7B mostly Q4_0           |   3.56 GiB | pp 512     |          4 |   1032.08 ± 0.60 |   1164.00 ± 1.86 |  1.128          |
| LLaMA 7B mostly Q4_1           |   3.95 GiB | pp 512     |          4 |   1038.05 ± 0.52 |   1161.63 ± 1.08 |  1.119          |
| LLaMA 7B mostly Q6_K           |   5.15 GiB | pp 512     |          4 |    856.86 ± 0.57 |    977.56 ± 0.74 |  1.141          |
| LLaMA 7B mostly Q5_K - Medium  |   4.45 GiB | pp 512     |          4 |    857.33 ± 0.36 |    976.31 ± 1.51 |  1.139          |
| LLaMA 7B mostly Q5_K - Small   |   4.33 GiB | pp 512     |          4 |    857.07 ± 0.45 |    977.60 ± 0.25 |  1.141          |
| LLaMA 7B mostly Q4_K - Medium  |   3.80 GiB | pp 512     |          4 |    920.57 ± 0.36 |   1028.20 ± 0.39 |  1.117          |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB | pp 512     |          4 |    930.13 ± 0.65 |   1034.84 ± 0.53 |  1.113          |
| LLaMA 7B mostly Q3_K - Medium  |   3.07 GiB | pp 512     |          4 |    903.22 ± 0.71 |   1008.01 ± 0.85 |  1.116          |
| LLaMA 7B mostly Q3_K - Small   |   2.75 GiB | pp 512     |          4 |    886.80 ± 0.67 |    990.63 ± 0.62 |  1.117          |
| LLaMA 7B mostly F16            |  12.55 GiB | tg 128     |          4 |     40.82 ± 0.02 |     40.93 ± 0.05 |  1.003          |
| LLaMA 7B mostly Q8_0           |   6.67 GiB | tg 128     |          4 |     64.50 ± 0.04 |     64.69 ± 0.04 |  1.003          |
| LLaMA 7B mostly Q4_0           |   3.56 GiB | tg 128     |          4 |     90.25 ± 0.11 |     91.04 ± 0.12 |  1.009          |
| LLaMA 7B mostly Q4_1           |   3.95 GiB | tg 128     |          4 |     85.47 ± 0.09 |     86.06 ± 0.08 |  1.007          |
| LLaMA 7B mostly Q6_K           |   5.15 GiB | tg 128     |          4 |     71.01 ± 0.08 |     71.51 ± 0.05 |  1.007          |
| LLaMA 7B mostly Q5_K - Medium  |   4.45 GiB | tg 128     |          4 |     72.36 ± 0.07 |     71.68 ± 0.06 |  0.991          |
| LLaMA 7B mostly Q5_K - Small   |   4.33 GiB | tg 128     |          4 |     73.75 ± 0.08 |     72.88 ± 0.06 |  0.988          |
| LLaMA 7B mostly Q4_K - Medium  |   3.80 GiB | tg 128     |          4 |     83.39 ± 0.13 |     83.70 ± 0.12 |  1.004          |
| LLaMA 7B mostly Q4_K - Small   |   3.59 GiB | tg 128     |          4 |     86.59 ± 0.11 |     86.94 ± 0.15 |  1.004          |
| LLaMA 7B mostly Q3_K - Medium  |   3.07 GiB | tg 128     |          4 |     83.50 ± 0.10 |     80.20 ± 0.08 |  0.960          |
| LLaMA 7B mostly Q3_K - Small   |   2.75 GiB | tg 128     |          4 |     84.73 ± 0.10 |     77.74 ± 0.06 |  0.918          |

Looks like there is a small performance hit on `Q3_K` TG with the Ultra. I redid the test 2 times, but the numbers are consistent.