@ikawrakow 2023-09-09
> Looks like there is a small performance hit on Q3_K TG with the Ultra.

I think this is because I did not rebase on latest master before opening the PR. In the meantime you merged #2995 into master, which brings significant improvement in TG for `Q3_K`. If you rebase the PR on current master, I expect the `Q3_K` performance drop to turn into a small performance increase.